Updates committee data and rendering code
The committee.yml sets up an associative array with the key being the name of the committee and the value is the data about the committee. This is needed in author guideline pages to easily display only the members of the committee for a specific track (see submission-questions.html)

It also updates the rendering of the committee to use a simpler table, though we may want to use a list eventualy to better incorporate any headshots for committee members.